+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Disable automatic tmux shell wrapping
 * Show DeepSeek balance inline
+* Copy on select and notify in foot
 
 ### 2.65.1: 2026-09-12